Update MockObservation to include MockObservationIterator and enhance test coverage
- Added MockObservationIterator.hpp to the gmock backend, facilitating iteration over mock observations.
- Updated MockObservation.hpp to utilize MockObservationIterator, improving code organization and clarity.
- Included MockObservation and MockObservationIterator in ObservationTest.cpp to enhance testing capabilities.
